исправил названия переменных и файлов
