+++ js/init-modal-popup.js
@@ -1,24 +1,24 @@
-import { userPosts } from './picture-render.js';
+import { posts } from './data.js';
 
-const picturesContainer = document.querySelector('.pictures');
-const fullPhoto = document.querySelector('.big-picture');
-const pictureCloseButton = document.querySelector('.big-picture__cancel');
-const commentsTemplate = document.querySelector('#comments').content.querySelector('.social__comment');
+const initModalPopup = () => {
+  const picturesContainer = document.querySelector('.pictures');
+  const fullPhoto = document.querySelector('.big-picture');
+  const pictureCloseButton = document.querySelector('.big-picture__cancel');
+  const commentsTemplate = document.querySelector('#comments').content.querySelector('.social__comment');
 
-const commentsContainer = document.querySelector('.social__comments');
-const bigPictureImg = document.querySelector('.big-picture__img img');
-const commentsCount = document.querySelector('.comments-count');
-const likesCount = document.querySelector('.likes-count');
-const socialCaption = document.querySelector('.social__caption');
-const commentCountArea = document.querySelector('.social__comment-count');
-const commentsLoader = document.querySelector('.comments-loader');
+  const commentsContainer = document.querySelector('.social__comments');
+  const bigPictureImg = document.querySelector('.big-picture__img img');
+  const commentsCount = document.querySelector('.comments-count');
+  const likesCount = document.querySelector('.likes-count');
+  const socialCaption = document.querySelector('.social__caption');
+  const commentCountArea = document.querySelector('.social__comment-count');
+  const commentsLoader = document.querySelector('.comments-loader');
 
-const similarCommentsFragment = document.createDocumentFragment();
+  const similarCommentsFragment = document.createDocumentFragment();
 
-const initPictures = () => {
   picturesContainer.addEventListener('click', (event) => {
     const postId = event.target.closest('a.picture').dataset.postId - 1;
-    const currentPost = userPosts[postId];
+    const currentPost = posts[postId];
     fullPhoto.classList.remove('hidden');
     document.body.classList.add('modal-open');
     commentCountArea.classList.add('hidden');
@@ -29,7 +29,7 @@ const initPictures = () => {
     socialCaption.textContent = currentPost.description;
     commentsContainer.innerHTML = '';
 
-    userPosts[postId].comments.forEach(({ avatar, message, name}) => {
+    posts[postId].comments.forEach(({ avatar, message, name}) => {
       const commentsItem = commentsTemplate.cloneNode(true);
       const commentImg = commentsItem.querySelector('.social__picture');
       commentImg.src = avatar;
@@ -53,4 +53,4 @@ const initPictures = () => {
   });
 };
 
-export { initPictures };
+export { initModalPopup };
