+++ js/main.js
@@ -1,3 +1,5 @@
-import { initPictures } from './picture-show.js';
+import { renderPictures } from './render-pictures.js';
+import { initModalPopup } from './init-modal-popup.js';
 
-initPictures();
+renderPictures();
+initModalPopup();
